refactoring btw
