+++ src/assets/assetload_funcs.cpp
@@ -1,4 +1,4 @@
-#include "asset_loaders.h"
+#include "assetload_funcs.h"
 
 #include <iostream>
 #include <filesystem>
