+++ src/assets/assetload_funcs.h
